+++ test/06-misc.js
@@ -1,6 +1,8 @@
-/* global expect, router */
+/* global expect, router, isNode */
 
 describe('Misc', function () {
+
+  var noop = function () {}
 
   this.beforeAll(function () {
     router.reset().listen()
@@ -11,17 +13,15 @@ describe('Misc', function () {
   })
 
   beforeEach(function () {
-    router.reset()
     expect.restoreSpies()
   })
 
   it('`clear` removes the routes leaving the context and callback', function () {
-    var noop = function () {}
     var done = false
     var path = '#/xyz'
     var data = { path: path, enter: noop }
 
-    router.add(data).onEnter(function () { done = true })
+    router.reset().add(data).onEnter(function () { done = true })
     router._run(path)
 
     var route = router.route(path)
@@ -47,10 +47,10 @@ describe('Misc', function () {
       { path: '#/abc' },
       { path: '#/xyz' }
     ]
-    var spy = expect.createSpy(function () {})
+    var spy = expect.createSpy(noop)
 
     location.hash = '#/'
-    router.add(data).onEnter(spy)
+    router.reset().add(data, noop).onEnter(spy).listen()
 
     var hash = data[0].path
     router.navigate(hash)
@@ -64,52 +64,61 @@ describe('Misc', function () {
     spy.reset()
 
     hash = data[1].path
-    router.navigate(hash)
+    router.onEnter(spy).listen().navigate(hash)
     expect(location.hash).toBe(hash)
     expect(spy).toHaveBeenCalled()
   })
 
   it('`navigate` can force the callback even without changes', function () {
     var hash = '#/abc'
-    var spy = expect.createSpy(function () {})
+    var spy = expect.createSpy(noop)
 
     location.hash = '#/'
-    router.add({ path: hash }, function () {}).onEnter(spy)
+    router.reset().add({ path: hash }, noop).listen().onEnter(spy)
 
     router.navigate(hash)
     expect(location.hash).toBe(hash)
-    expect(spy.calls.length).toBe(1)
+    isNode && expect(spy.calls.length).toBe(1)
 
     router.navigate(hash, true)
     expect(location.hash).toBe(hash)
-    expect(spy.calls.length).toBe(2)
+    isNode && expect(spy.calls.length).toBe(2)
 
     router.navigate(hash, true)
     expect(location.hash).toBe(hash)
-    expect(spy.calls.length).toBe(3)
+    isNode && expect(spy.calls.length).toBe(3)
   })
 
-  it('`getContext` will retrieve the current route context', function () {
+  it('`getContext` will retrieve the current router context', function () {
     var path = '#/abc/:id'
     var hash = '#/abc/1'
-    router.reset().add({ path: path }).navigate(hash, true)
 
-    var route = router.getContext()
-    expect(route).toBeAn('object').toExist()
-    expect(route).toInclude({ path: path, hash: hash })
-    expect(route.params).toExist().toInclude({ id: '1' })
+    router.reset().add({ path: path }, noop).listen().navigate(hash, true)
+
+    if (isNode) {
+      var ctx = router.getContext()
+
+      expect(ctx).toBeAn('object').toExist()
+      expect(ctx).toInclude({ lastHash: hash, onEnter: undefined })
+      expect(ctx.lastRoute).toExist().toInclude({ path: path, hash: hash })
+      expect(ctx.lastRoute.params).toExist().toInclude({ id: '1' })
+    }
   })
 
-  it('`getContext` will return `null` after a reset', function () {
+  it('`getContext` lastRoute will be `null` after a reset', function () {
     var path = '#/abc'
-    router.reset().add({ path: path }).navigate(path, true)
+    router.reset().add({ path: path }, noop).listen().navigate(path, true)
 
-    var route = router.getContext()
-    expect(route).toBeAn('object').toExist()
+    if (isNode) {
+      var ctx = router.getContext()
+      expect(ctx).toBeAn('object').toExist()
+      expect(ctx.lastRoute).toExist().toInclude({ hash: path })
 
-    router.reset()
-    route = router.getContext()
-    expect(route).toBe(null)
+      router.reset()
+      ctx = router.getContext()
+      expect(ctx).toBeAn('object').toExist()
+      expect(ctx.lastRoute).toBe(null)
+    }
   })
 
 })