+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Global callbacks `onEnter` and `onExit`.
 - TypeScript v3 typings.
 - Test for 100% coverage.
+- Browser test with karma (local).
+- Method `getContext` to retrieve the router context.
 
 ### Changed